Large init, logic and generate refactor
* Constant time searching via map[string]
* Share code between interfacer and generator via internal pkg
* Generate full interface and func types, as well as names
* The above avoids having to import all of std at startup

For now, we are using the string represenations of interfaces and funcs
to index them. Since these strings are bounded to a certain amount of
characters, using hashes is not important.

There is only one drawback to using strings - you can only find exact
matches. If later on we want to find partial matches, like an interface
that satisfies our var usage but has an extra func, we will need to come
up with some data structure changes to do that search efficiently.