Rewrite package to allow better use as a library
The only way to use the package before was via an io.Writer, which was
intended for command line output. This is still an option, but now
another lower level function is available to receive the warnings
themselves and be able to do things with them other than write them as
strings somewhere.