Only return results per package, not per warning
The ideal check is per package, not per file nor per function. For
example, if we check an entire package we know for sure whether an
unexported function is implementing an interface or not. We can't do
that for exported functions as they might be used outside of the
package.