Named types without methods are not interesting
I noticed this in the last commit when go/types.ImportMode was added as
an interesting parameter type. It isn't, since it's just a named type
with zero methods. We would never suggest that such a parameter be an
interface, since we can never call any methods on it.

This also removes the dead code path for the types.Struct case. Struct
literals as parameter types are not interesting either.

Run go generate again.