Avoid false positives when assigning functions
Basically, whenever we use a function as a value we know that we are
discarding the function in question. Be it because we're assigning it to
some func variable, or passing it as a func parameter, the func
signature on the receiving end must be exactly the same.

Thus, if we're using a func as a value, its signature cannot change
without breaking compilation (unless more changes are done).

We are doing this in a hacky way, by using the *types.Signature object.
Even if two funcs have equivalent signatures, their signatures are
actually not the same pointers. This lets us identify which func we are
discarding independently of whether it was assigned to an identifier, if
we've seen it yet, etc.

Fixes #20.