Don't inspect function literals
They're often declared in a local scope and will never be reused or
exposed. Warnings on these weren't false positives, but were still
annoying.

Moreover, the `return nil` there was potentially introducing false
positives as we were discarding bodies of nested functions, which could
be using parameters of the parent function.