Do typecheck std func bodies if requested
Otherwise, `interfacer std` would not actually do any typechecking and
we would get calls having nil types, rendering the linter useless or
making it crash.

Always typechecking std is wasteful; If the linter is run against a
non-std package, typechecking all of std is not actually useful and
wastes a lot of time. For example, with and without std typechecking:

	 $ time interfacer github.com/mvdan/sh/...
	real    0m1.102s
	user    0m2.413s
	sys     0m0.083s
	 $ time interfacer github.com/mvdan/sh/...
	real    0m0.383s
	user    0m1.270s
	sys     0m0.057s

For now, be conservative and simply add typechecking of the std packages
that are specified as arguments.

We should somehow be able to only type check the packages specified as
arguments, but that doesn't work with `interfacer foo.go` as at this
point we don't yet know what package foo.go belongs to. We know once we
call Load(), but that's too late to set up TypeCheckFuncBodies.

Also add a test.

Fixes #28.