Stop generating code for std types
Instead, suggest only those types available in the package itself and
its direct dependencies. This makes the tool simpler and it will also
suggest closer types.

This also means that, for example, that if "io" or "fmt" aren't
imported, their interfaces won't be suggested. But this can be a good
thing - most of the time, these suggestions wouldn't be particularly
useful.