README: formally deprecate the tool
I had stopped actively working on it for a while for the reason
explained below. Let's make it official, to ensure that the tool is no
longer recommended for everyday use.

If the tool were to be continued, it would also need to be rewritten to
use x/tools/go/ssa. This is the first static analysis tool I ever wrote,
so it uses the much more rudimentary go/ast, limiting its capabilities.